Fix failing test in order to check if adding a review sets a new id

+++ src/main/java/com/teamtreehouse/techdegrees/dao/TodoDao.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface TodoDao {
     void add(Todo todo) throws DaoException;
     List<Todo> findAll();
-    Todo findById(long id);
+    Todo findById(int id);
 }
